requirements.txt
Django==5.0.1
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-dotenv==1.0.0
certifi==2024.2.2
urllib3==2.2.0
requests==2.31.0
google-generativeai==0.3.2
redis==5.0.1
django-redis==5.4.0
gunicorn==21.2.0
whitenoise==6.6.0
sentry-sdk==1.40.0
PyJWT==2.8.0